+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [0.9.5] - UNRELEASED
 
+- Added log message when creating zip output
 - Fixed disable pre-patched materials patcher wrongly disabling complex materials with dynamic cubemaps
+- Fixed slow output zip creation
 
 ## [0.9.4] - 2025-10-19